refactor(registry): move block side menu into notion example
