+++ registry/package.json
@@ -24,7 +24,6 @@
     "./lit/examples/toolbar": "./src/lit/examples/toolbar/index.ts",
     "./lit/renderer": "./src/lit/renderer.ts",
     "./preact/examples/block-handle": "./src/preact/examples/block-handle/index.ts",
-    "./preact/examples/block-side-menu": "./src/preact/examples/block-side-menu/index.ts",
     "./preact/examples/blockquote": "./src/preact/examples/blockquote/index.ts",
     "./preact/examples/bold": "./src/preact/examples/bold/index.ts",
     "./preact/examples/change-tracking": "./src/preact/examples/change-tracking/index.ts",
@@ -76,7 +75,6 @@
     "./preact/examples/yjs": "./src/preact/examples/yjs/index.ts",
     "./preact/renderer": "./src/preact/renderer.ts",
     "./react/examples/block-handle": "./src/react/examples/block-handle/index.ts",
-    "./react/examples/block-side-menu": "./src/react/examples/block-side-menu/index.ts",
     "./react/examples/blockquote": "./src/react/examples/blockquote/index.ts",
     "./react/examples/bold": "./src/react/examples/bold/index.ts",
     "./react/examples/change-tracking": "./src/react/examples/change-tracking/index.ts",
@@ -132,7 +130,6 @@
     "./react/renderer": "./src/react/renderer.ts",
     "./replace-class-names": "./src/meta/replace-classes.ts",
     "./solid/examples/block-handle": "./src/solid/examples/block-handle/index.ts",
-    "./solid/examples/block-side-menu": "./src/solid/examples/block-side-menu/index.ts",
     "./solid/examples/blockquote": "./src/solid/examples/blockquote/index.ts",
     "./solid/examples/bold": "./src/solid/examples/bold/index.ts",
     "./solid/examples/change-tracking": "./src/solid/examples/change-tracking/index.ts",
@@ -184,7 +181,6 @@
     "./solid/examples/yjs": "./src/solid/examples/yjs/index.ts",
     "./solid/renderer": "./src/solid/renderer.ts",
     "./svelte/examples/block-handle": "./src/svelte/examples/block-handle/index.ts",
-    "./svelte/examples/block-side-menu": "./src/svelte/examples/block-side-menu/index.ts",
     "./svelte/examples/blockquote": "./src/svelte/examples/blockquote/index.ts",
     "./svelte/examples/bold": "./src/svelte/examples/bold/index.ts",
     "./svelte/examples/change-tracking": "./src/svelte/examples/change-tracking/index.ts",
@@ -241,7 +237,6 @@
     "./vanilla/renderer": "./src/vanilla/renderer.ts",
     "./vite-plugin-class-replace": "./src/meta/vite-plugin-class-replace.ts",
     "./vue/examples/block-handle": "./src/vue/examples/block-handle/index.ts",
-    "./vue/examples/block-side-menu": "./src/vue/examples/block-side-menu/index.ts",
     "./vue/examples/blockquote": "./src/vue/examples/blockquote/index.ts",
     "./vue/examples/bold": "./src/vue/examples/bold/index.ts",
     "./vue/examples/change-tracking": "./src/vue/examples/change-tracking/index.ts",
