--- registry/src/preact/examples/block-side-menu/index.ts
@@ -1,1 +1,0 @@
-export { default as ExampleEditor } from './editor.tsx'
