--- registry/src/react/examples/block-side-menu/editor.tsx
@@ -1,38 +1,0 @@
-'use client'
-
-import 'prosekit/basic/style.css'
-import 'prosekit/basic/typography.css'
-
-import { createEditor, type NodeJSON } from 'prosekit/core'
-import { ProseKit } from 'prosekit/react'
-import { useMemo } from 'react'
-
-import { sampleContent } from '../../sample/sample-doc-block-handle.ts'
-import { DropIndicator } from '../../ui/drop-indicator/index.ts'
-import { defineExtension } from '../block-handle/extension.ts'
-
-import BlockSideMenu from './block-side-menu.tsx'
-
-interface EditorProps {
-  initialContent?: NodeJSON
-}
-
-export default function Editor(props: EditorProps) {
-  const defaultContent = props.initialContent ?? sampleContent
-  const editor = useMemo(() => {
-    const extension = defineExtension()
-    return createEditor({ extension, defaultContent })
-  }, [defaultContent])
-
-  return (
-    <ProseKit editor={editor}>
-      <div className="CSS_EDITOR_VIEWPORT">
-        <div className="CSS_EDITOR_SCROLLING">
-          <div ref={editor.mount} className="CSS_EDITOR_CONTENT"></div>
-          <BlockSideMenu />
-          <DropIndicator />
-        </div>
-      </div>
-    </ProseKit>
-  )
-}
