+++ website/src/content/docs/components/block-handle.mdx
@@ -113,7 +113,7 @@ function handleStateChange(event) {
 
 Use the built-in `BlockHandleDraggable` when you only need drag-and-drop. If the same handle also opens a framework menu, implement that draggable trigger in your framework layer so the click, keyboard, menu open state, and menu actions all live in the same reactive system.
 
-See the `block-side-menu` example for a framework-rendered draggable handle that also opens a side menu and runs actions like changing the block to `Heading 1` or deleting the block.
+See the Notion example for a framework-rendered draggable handle that also opens a side menu and runs actions like changing the block type or deleting the block.
 
 ## API reference
